Worked around the color cycling problem and loaded demo for the office
I worked around the color cycling problem but I'm not sure if I really
know what's going on. There seemed to be something funky with the
NeoMatrix library and colors, although there's a chance it is because I
had the whole matrix brightness turned down to 20 (out of 255). Looking
in the code, there may be funky stuff from using matrix brightness. I
switched all the color cycling stuff to use the legacy API and it works
much better.

There is a bug that is kind of cool though. It looks like the legacy API
and the Matrix API use two different buffers so clearing one doesn't
clear the other. This can be seen where the Yellow Rectangle and the 3
X's show up on a light colored background.